backend/requirements/base.txt
Django==6.0.2
djangorestframework==3.15.2
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
celery==5.4.0
django-celery-results==2.5.1
redis==5.2.0
mysqlclient==2.2.6
Pillow==11.1.0
openpyxl==3.1.5
reportlab==4.4.4
pypdf==5.4.0
gunicorn==23.0.0
python-decouple==3.8
bcrypt==4.2.1
